splitting components into smaller components
